refactor: move keymap legends to its own file
To reduce the length of editor.rs

TODO:
- remove fsl and friends
