+++ src/components/mod.rs
@@ -3,6 +3,7 @@ pub mod dropdown;
 pub mod prompt;
 
 pub mod editor;
+mod editor_keymap_legend;
 pub mod file_explorer;
 pub mod keymap_legend;
 pub mod suggestive_editor;
